added chirp object and modified timestamp type in AddChirp

+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/model/network/method/message/data/socialmedia/AddChirp.java
@@ -19,7 +19,7 @@ public class AddChirp extends Data {
   @Nullable
   private final String parentId;
 
-  private final long timestamp;
+  private final Long timestamp;
 
   /**
    * Constructor for a data Add Chirp
@@ -28,7 +28,7 @@ public class AddChirp extends Data {
    * @param parentId message ID of parent chirp, can be null
    * @param timestamp UNIX timestamp in UTC
    */
-  public AddChirp(String text, @Nullable String parentId, int timestamp) {
+  public AddChirp(String text, @Nullable String parentId, Long timestamp) {
     if (text.length() > MAX_CHIRP_CHARS) {
       throw new IllegalArgumentException("the text exceed the maximum numbers of characters");
     }
